requirements.txt
discord.py == 2.2.2
python_dotenv >= 0.20.0
pytz >= 2022.5
beautifulsoup4
requests